refactor: mover classes provisórias de geralanding para subpacotes
