+++ backend/ads-service/src/test/java/com/marketinghub/geralanding/WireframeHtmlGeneratorTest.java
@@ -1,6 +1,6 @@
 package com.marketinghub.geralanding;
 
-import com.marketinghub.geralanding.wireframe.WireframeHtmlGenerator;
+import com.marketinghub.geralanding.wireframe.provisorio.WireframeHtmlGenerator;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
