+++ backend/ads-service/src/test/java/com/marketinghub/geralanding/WireframeProvisionalHtmlAssemblerTest.java
@@ -1,7 +1,7 @@
 package com.marketinghub.geralanding;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.marketinghub.geralanding.wireframe.WireframeProvisionalHtmlAssembler;
+import com.marketinghub.geralanding.wireframe.provisorio.WireframeProvisionalHtmlAssembler;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
